+++ src/features/audio-file/primitives/audio-files-manager.ts
@@ -1,0 +1,67 @@
+import { err, ok, type ResultAsync } from "neverthrow";
+import { type Accessor, createSignal } from "solid-js";
+import { createStore, reconcile } from "solid-js/store";
+import type { UpdateAudioFile } from "@/tauri/commands";
+import type { AudioFileDTO, AudioFilePatchDTO } from "@/tauri/dto";
+
+export interface AudioFilesManager {
+  audioFiles: AudioFileDTO[];
+  isUpdating: Accessor<boolean>;
+
+  /** 音声ファイルを更新する
+   *
+   * 更新を永続化する
+   */
+  updateAudioFile: (
+    patch: AudioFilePatchDTO,
+  ) => ResultAsync<AudioFileDTO, string>;
+  /**
+   * オーディオファイルリストを新しいリストに置き換える
+   *
+   * 現在管理しているファイルをすべてクリアし、新しいファイルリストで置き換える
+   */
+  replaceAllAudioFiles: (audioFiles: AudioFileDTO[]) => void;
+}
+
+export const createAudioFilesManager = (
+  initialAudioFiles: AudioFileDTO[],
+  { updateAudioFileCommand }: { updateAudioFileCommand: UpdateAudioFile },
+): AudioFilesManager => {
+  const [audioFiles, setAudioFiles] = createStore(initialAudioFiles);
+  const [isUpdating, setIsUpdating] = createSignal<boolean>(false);
+
+  const updateAudioFile = (
+    patch: AudioFilePatchDTO,
+  ): ResultAsync<AudioFileDTO, string> => {
+    setIsUpdating(true);
+
+    return updateAudioFileCommand({ patch })
+      .andThen((newFile) => {
+        const fileIndex = audioFiles.findIndex(
+          (file) => file.id === newFile.id,
+        );
+        if (fileIndex === -1) {
+          return err("File not found");
+        }
+
+        setAudioFiles(fileIndex, reconcile(newFile));
+        setIsUpdating(false);
+        return ok(newFile);
+      })
+      .orElse((e) => {
+        setIsUpdating(false);
+        return err(e);
+      });
+  };
+
+  const replaceAllAudioFiles = (audioFiles: AudioFileDTO[]) => {
+    setAudioFiles(audioFiles);
+  };
+
+  return {
+    audioFiles,
+    isUpdating,
+    updateAudioFile,
+    replaceAllAudioFiles,
+  };
+};
